fixed failed tests and deleted unnecessary entity
