+++ question-service/src/main/java/com/information/center/questionservice/controller/QuestionController.java
@@ -6,8 +6,10 @@ import com.information.center.questionservice.model.request.QuestionRequest;
 import com.information.center.questionservice.model.request.QuestionRequestValidation;
 import com.information.center.questionservice.model.response.QuestionResponse;
 import com.information.center.questionservice.model.response.QuestionResponsePage;
+import com.information.center.questionservice.service.QuestionService;
 import com.information.center.questionservice.service.QuestionServiceImpl;
 import com.information.center.questionservice.service.QuestionValidateService;
+import com.information.center.questionservice.service.QuestionValidateServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -18,50 +20,43 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class QuestionController implements QuestionEndpoint {
 
-    private final QuestionServiceImpl questionService;
+    private final QuestionService questionService;
 
     private final QuestionValidateService questionValidateService;
 
-    @Override
     @PostMapping("/topic/{topicExternalId}")
     public QuestionResponse create(@RequestBody QuestionRequest questionRequest, @PathVariable("topicExternalId") String topicExternalId) {
 
         return questionService.create(questionRequest, topicExternalId);
     }
 
-    @Override
     @GetMapping("/questionsByTopic/{topicExternalId}")
     public QuestionListDetails findQuestionsByTopicId(@PathVariable("topicExternalId") String topicExternalId,
                                                       Pageable pageable) {
         return questionService.findQuestionsByTopicId(topicExternalId, pageable);
     }
 
-    @Override
     @PutMapping
     public void update(@RequestBody QuestionRequest questionRequest) {
         questionService. update(questionRequest);
     }
 
-    @Override
     @GetMapping("/{externalId}")
     public QuestionResponse findByExternalId(@PathVariable("externalId") String externalId) {
         return questionService.findByExternalId(externalId);
     }
 
-    @Override
     @GetMapping
     public QuestionResponsePage findAll(Pageable pageable) {
         return questionService.findAll(pageable);
     }
 
-    @Override
     @DeleteMapping("/{externalId}")
     public void delete(@PathVariable("externalId") String externalId) {
 
         questionService.delete(externalId);
     }
 
-    @Override
     @PostMapping("/validate")
     public QuestionResponseValidated validate(@RequestBody QuestionRequestValidation questionRequestValidation) {
         return questionValidateService.validate(questionRequestValidation);
