+++ question-service/src/main/java/com/information/center/questionservice/model/QuestionResponseValidated.java
@@ -2,6 +2,7 @@ package com.information.center.questionservice.model;
 
 import com.information.center.questionservice.model.response.QuestionResponse;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -11,6 +12,7 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@Builder
 public class QuestionResponseValidated {
 
     private Date startDate;
